+++ tfexec/internal/testutil/tfcache.go
@@ -1,11 +1,6 @@
 package testutil
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
 	"sync"
 	"testing"
 
@@ -48,43 +43,4 @@ func (tf *TFCache) Version(t *testing.T, v string) string {
 		f.UserAgent = appendUserAgent
 		return f
 	})
-}
-
-func (tf *TFCache) find(t *testing.T, key string, finder func(dir string) tfinstall.ExecPathFinder) string {
-	t.Helper()
-
-	if tf.dir == "" {
-		// panic instead of t.fatal as this is going to affect all downstream tests reusing the cache entry
-		panic("installDir not yet configured")
-	}
-
-	tf.Lock()
-	defer tf.Unlock()
-
-	if path, ok := tf.execs[key]; ok {
-		return path
-	}
-
-	keyDir := key
-	keyDir = strings.ReplaceAll(keyDir, ":", "-")
-	keyDir = strings.ReplaceAll(keyDir, "/", "-")
-
-	dir := filepath.Join(tf.dir, keyDir)
-
-	t.Logf("caching exec %q in dir %q", key, dir)
-
-	err := os.MkdirAll(dir, 0777)
-	if err != nil {
-		// panic instead of t.fatal as this is going to affect all downstream tests reusing the cache entry
-		panic(fmt.Sprintf("unable to mkdir %q: %s", dir, err))
-	}
-
-	path, err := tfinstall.Find(context.Background(), finder(dir))
-	if err != nil {
-		// panic instead of t.fatal as this is going to affect all downstream tests reusing the cache entry
-		panic(fmt.Sprintf("error installing terraform %q: %s", key, err))
-	}
-	tf.execs[key] = path
-
-	return path
 }
