+++ tfexec/show_test.go
@@ -2,7 +2,6 @@ package tfexec
 
 import (
 	"context"
-	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-exec/tfexec/internal/testutil"
@@ -10,7 +9,6 @@ import (
 
 func TestShowCmd(t *testing.T) {
 	td := testTempDir(t)
-	defer os.RemoveAll(td)
 
 	tf, err := NewTerraform(td, tfVersion(t, testutil.Latest012))
 	if err != nil {
@@ -32,7 +30,6 @@ func TestShowCmd(t *testing.T) {
 
 func TestShowStateFileCmd(t *testing.T) {
 	td := testTempDir(t)
-	defer os.RemoveAll(td)
 
 	tf, err := NewTerraform(td, tfVersion(t, testutil.Latest012))
 	if err != nil {
@@ -54,7 +51,6 @@ func TestShowStateFileCmd(t *testing.T) {
 
 func TestShowPlanFileCmd(t *testing.T) {
 	td := testTempDir(t)
-	defer os.RemoveAll(td)
 
 	tf, err := NewTerraform(td, tfVersion(t, testutil.Latest012))
 	if err != nil {
@@ -76,7 +72,6 @@ func TestShowPlanFileCmd(t *testing.T) {
 
 func TestShowPlanFileRawCmd(t *testing.T) {
 	td := testTempDir(t)
-	defer os.RemoveAll(td)
 
 	tf, err := NewTerraform(td, tfVersion(t, testutil.Latest012))
 	if err != nil {
